MarkItUp! editor for forum posts.
New bbcode [right] added. Should work the same way as center. Added css for both classes.
The left button currently clears right and center tags in the selection. This is not enough if you want to left align in the middle of a center block. We could:

 * Implement a more advanced logic including a [left] tag. This can lead to unnecessary heaps of tags.
 * Remove the button, no need for a feature that doesn't do what it suggests properly.
 * Leave it as-is.

A colorpicker library is also included. This library is quite unpleasant to customise, but the defaults work decently and it looks reasonably. I'm not sure it will work with multiple editors on a page though (it's currently not an issue).

Icons are mostly famfamfam, except for the ugly ones. They reflect my paint skills.