Revert "Check CSRF on unsubscribe, too."
This reverts commit e437ce62085b755c45c61953c88003eccb78660b.

Checking CSRF on unsubscribe means that unsubscribe links in emails can't work. This leaves a bit of a hole open but it's not especially destructive, at least. There are ways to make it better, but it'd be a bit more involved than this... something like ?UID=x&SomeUserHash=y would do the job decently, I think, it'd require anything exploiting this to be specifically targeted to a user who's had that bit of info stolen somehow, but if you've got that info you've probably got their email account, and you've got bigger problems.